+++ crates/secrets/src/bin/keeper.rs
@@ -9,15 +9,21 @@ use tokio::net::UnixListener;
 use tokio::signal;
 use tokio::task::JoinHandle;
 
-// Test constants moved to test section
-
-// Additional test constants
+// Test constants
 #[cfg(test)]
 const PROMPT_NO_VAULT_FOUND: &str = "no vault found";
 #[cfg(test)]
 const PROMPT_VAULT_CREATED: &str = "vault created successfully";
 #[cfg(test)]
 const PROMPT_DAEMON_STARTED: &str = "daemon started";
+#[cfg(test)]
+const PROMPT_ENTER_NEW_PASSWORD: &str = "enter new master password:";
+#[cfg(test)]
+const PROMPT_CONFIRM_PASSWORD: &str = "confirm master password:";
+#[cfg(test)]
+const ERROR_PASSWORD_EMPTY: &str = "master password cannot be empty";
+#[cfg(test)]
+const ERROR_PASSWORDS_DONT_MATCH: &str = "passwords do not match";
 #[cfg(test)]
 const ERROR_INCORRECT_PASSWORD: &str = "incorrect password";
 
@@ -982,197 +988,6 @@ mod tests {
   }
 
   #[test]
-  fn test_verify_password_success() {
-    use secrets::PasswordBasedCredentialStore;
-    use std::collections::HashMap;
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let vault_path = temp_dir.path().join("test_vault.enc");
-    let test_password = "test_verification_password_123";
-
-    // Create a valid vault file
-    let empty_credentials = HashMap::new();
-    let store = PasswordBasedCredentialStore::new(&empty_credentials, test_password).unwrap();
-    store.save_to_file(&vault_path).unwrap();
-
-    // Test successful password verification
-    let result = verify_password(&vault_path, test_password);
-    assert!(result.is_ok(), "Password verification should succeed with correct password");
-  }
-
-  #[test]
-  fn test_verify_password_incorrect_password() {
-    use secrets::PasswordBasedCredentialStore;
-    use std::collections::HashMap;
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let vault_path = temp_dir.path().join("test_vault.enc");
-    let correct_password = "correct_password_456";
-    let wrong_password = "wrong_password_789";
-
-    // Create a valid vault file with correct password
-    let empty_credentials = HashMap::new();
-    let store = PasswordBasedCredentialStore::new(&empty_credentials, correct_password).unwrap();
-    store.save_to_file(&vault_path).unwrap();
-
-    // Test password verification failure
-    let result = verify_password(&vault_path, wrong_password);
-    assert!(result.is_err(), "Password verification should fail with incorrect password");
-
-    let error_msg = result.unwrap_err().to_string();
-    assert!(
-      error_msg.contains("incorrect password"),
-      "Error should mention incorrect password, got: {error_msg}"
-    );
-  }
-
-  #[test]
-  fn test_verify_password_invalid_vault_format() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let vault_path = temp_dir.path().join("invalid_vault.enc");
-    let test_password = "any_password";
-
-    // Create a file with invalid JSON format (missing encrypted_data field)
-    let invalid_json = r#"{"some_other_field": "value"}"#;
-    std::fs::write(&vault_path, invalid_json).unwrap();
-
-    // Test that invalid vault format is detected
-    let result = verify_password(&vault_path, test_password);
-    assert!(result.is_err(), "Should fail with invalid vault format");
-
-    let error_msg = result.unwrap_err().to_string();
-    assert!(
-      error_msg.contains("invalid vault format") && error_msg.contains("encrypted_data"),
-      "Error should mention invalid vault format and missing encrypted_data, got: {error_msg}"
-    );
-  }
-
-  #[test]
-  fn test_verify_password_malformed_json() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let vault_path = temp_dir.path().join("malformed_vault.enc");
-    let test_password = "any_password";
-
-    // Create a file with completely malformed JSON
-    std::fs::write(&vault_path, "not json at all").unwrap();
-
-    // Test that malformed JSON is handled
-    let result = verify_password(&vault_path, test_password);
-    assert!(result.is_err(), "Should fail with malformed JSON");
-
-    // The exact error message will depend on serde_json, but it should fail
-    let _error_msg = result.unwrap_err().to_string();
-  }
-
-  #[test]
-  fn test_verify_password_file_not_found() {
-    use std::path::PathBuf;
-
-    let nonexistent_path = PathBuf::from("/tmp/definitely_does_not_exist.enc");
-    let test_password = "any_password";
-
-    // Test that missing file is handled
-    let result = verify_password(&nonexistent_path, test_password);
-    assert!(result.is_err(), "Should fail when vault file doesn't exist");
-
-    // Should get a file system error
-    let _error_msg = result.unwrap_err().to_string();
-  }
-
-  #[test]
-  fn test_create_new_vault_directory_creation() {
-    use secrets::PasswordBasedCredentialStore;
-    use std::collections::HashMap;
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let nested_path = temp_dir.path().join("deep").join("nested").join("path").join("vault.enc");
-    let test_password = "directory_test_password_123";
-
-    // Verify the parent directory doesn't exist initially
-    assert!(!nested_path.parent().unwrap().exists());
-
-    // Create a vault directly to test the directory creation logic
-    let empty_credentials = HashMap::new();
-    let store = PasswordBasedCredentialStore::new(&empty_credentials, test_password).unwrap();
-
-    // This should create the parent directories
-    let result = store.save_to_file(&nested_path);
-    assert!(result.is_ok(), "Should be able to save to nested path");
-
-    // Verify parent directories were created
-    assert!(nested_path.parent().unwrap().exists(), "Parent directories should be created");
-    assert!(nested_path.exists(), "Vault file should be created");
-  }
-
-  #[test]
-  fn test_create_new_vault_store_creation_and_save() {
-    use secrets::PasswordBasedCredentialStore;
-    use std::collections::HashMap;
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let vault_path = temp_dir.path().join("test_store_vault.enc");
-    let test_password = "store_creation_password_456";
-
-    // Test the store creation and save logic that's used in create_new_vault
-    let empty_credentials = HashMap::new();
-    let store = PasswordBasedCredentialStore::new(&empty_credentials, test_password).unwrap();
-
-    // Test saving to file
-    let save_result = store.save_to_file(&vault_path);
-    assert!(save_result.is_ok(), "Should be able to save store to file");
-
-    // Verify the file was created and is valid
-    assert!(vault_path.exists(), "Vault file should exist after saving");
-
-    // Verify we can read it back and verify with the password
-    let verify_result = verify_password(&vault_path, test_password);
-    assert!(verify_result.is_ok(), "Should be able to verify the created vault");
-  }
-
-  #[test]
-  fn test_create_new_vault_store_with_trimmed_password() {
-    use secrets::PasswordBasedCredentialStore;
-    use std::collections::HashMap;
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let vault_path = temp_dir.path().join("trimmed_password_vault.enc");
-    let password_with_whitespace = "  password_with_spaces  ";
-    let trimmed_password = "password_with_spaces";
-
-    // Test that store creation trims the password (like create_new_vault does)
-    let empty_credentials = HashMap::new();
-    let store =
-      PasswordBasedCredentialStore::new(&empty_credentials, password_with_whitespace.trim())
-        .unwrap();
-
-    store.save_to_file(&vault_path).unwrap();
-
-    // Should verify with the trimmed version
-    let verify_result = verify_password(&vault_path, trimmed_password);
-    assert!(verify_result.is_ok(), "Should verify with trimmed password");
-
-    // Should ALSO verify with the untrimmed version because verify_password also trims
-    let verify_untrimmed = verify_password(&vault_path, password_with_whitespace);
-    assert!(
-      verify_untrimmed.is_ok(),
-      "Should also verify with untrimmed password because verify_password trims it"
-    );
-
-    // But should NOT verify with a completely different password
-    let verify_wrong = verify_password(&vault_path, "completely_wrong_password");
-    assert!(verify_wrong.is_err(), "Should not verify with wrong password");
-  }
-
-  #[test]
   fn test_prompt_for_password_function_exists() {
     // This test just verifies that prompt_for_password function compiles and can be called
     // We can't easily test its interactive behavior, but we can ensure it exists
@@ -1655,279 +1470,7 @@ mod tests {
     assert!(matches!(not_found.kind(), io::ErrorKind::NotFound));
   }
 
-  // Helper function to test create_new_vault logic without interactive prompts
-  fn create_new_vault_non_interactive(
-    cred_path: &Path,
-    password1: &str,
-    password2: &str,
-  ) -> Result<String> {
-    bentley::info("no vault found. creating new vault...");
+  // create_new_vault_non_interactive helper moved to crate::encryption::tests
 
-    // Simulate the empty password check (line 108-109)
-    if password1.trim().is_empty() {
-      return Err(anyhow!(ERROR_PASSWORD_EMPTY));
-    }
-
-    // Simulate the password mismatch check (line 113-114)
-    if password1 != password2 {
-      return Err(anyhow!(ERROR_PASSWORDS_DONT_MATCH));
-    }
-
-    // Test the HashMap creation (line 117)
-    let empty_credentials = std::collections::HashMap::new();
-
-    // Test the use import (line 118)
-    use secrets::PasswordBasedCredentialStore;
-
-    // Test the store creation (line 119)
-    let store = PasswordBasedCredentialStore::new(&empty_credentials, password1.trim())?;
-
-    // Test the directory creation logic (lines 121-123)
-    if let Some(parent) = cred_path.parent() {
-      fs::create_dir_all(parent)?;
-    }
-
-    // Test the save operation (line 125)
-    store.save_to_file(&cred_path.to_path_buf())?;
-
-    // Test the success logging (line 127)
-    bentley::success("vault created successfully");
-
-    // Test the return with trimming (line 128)
-    Ok(password1.trim().to_string())
-  }
-
-  #[test]
-  fn test_create_new_vault_empty_password_path() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Test the empty password check (lines 108-109)
-    let result = create_new_vault_non_interactive(&cred_path, "", "anything");
-    assert!(result.is_err());
-    let error_msg = result.unwrap_err().to_string();
-    assert!(error_msg.contains("master password cannot be empty"));
-
-    // Test whitespace-only password
-    let result2 = create_new_vault_non_interactive(&cred_path, "   ", "   ");
-    assert!(result2.is_err());
-    let error_msg2 = result2.unwrap_err().to_string();
-    assert!(error_msg2.contains("master password cannot be empty"));
-  }
-
-  #[test]
-  fn test_create_new_vault_password_mismatch_path() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Test the password mismatch check (lines 113-114)
-    let result = create_new_vault_non_interactive(&cred_path, "password123", "different_password");
-    assert!(result.is_err());
-    let error_msg = result.unwrap_err().to_string();
-    assert!(error_msg.contains("passwords do not match"));
-  }
-
-  #[test]
-  fn test_create_new_vault_success_path_all_components() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let nested_path = temp_dir.path().join("nested").join("path");
-    let cred_path = nested_path.join("credentials.enc");
-
-    // Test the successful flow covering all lines
-    let result =
-      create_new_vault_non_interactive(&cred_path, "test_password_123", "test_password_123");
-    assert!(result.is_ok());
-    let returned_password = result.unwrap();
-    assert_eq!(returned_password, "test_password_123");
-
-    // Verify the directory was created (lines 121-123)
-    assert!(nested_path.exists());
-    assert!(nested_path.is_dir());
-
-    // Verify the credentials file was created (line 125)
-    assert!(cred_path.exists());
-
-    // Verify the file can be read back
-    let verify_result = verify_password(&cred_path, "test_password_123");
-    assert!(verify_result.is_ok());
-  }
-
-  #[test]
-  fn test_create_new_vault_trimming_in_success_path() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Test password trimming in success path (line 128)
-    let password_with_spaces = "  trimmed_password  ";
-    let result =
-      create_new_vault_non_interactive(&cred_path, password_with_spaces, password_with_spaces);
-    assert!(result.is_ok());
-    let returned_password = result.unwrap();
-
-    // Should be trimmed in return (line 128)
-    assert_eq!(returned_password, "trimmed_password");
-    assert!(!returned_password.contains(' '));
-
-    // Should be able to verify with trimmed password
-    let verify_result = verify_password(&cred_path, "trimmed_password");
-    assert!(verify_result.is_ok());
-  }
-
-  #[test]
-  fn test_create_new_vault_directory_creation_edge_case() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Test when parent directory already exists (lines 121-123)
-    assert!(cred_path.parent().unwrap().exists()); // Parent should already exist
-
-    let result =
-      create_new_vault_non_interactive(&cred_path, "existing_dir_test", "existing_dir_test");
-    assert!(result.is_ok());
-    assert!(cred_path.exists());
-  }
-
-  #[test]
-  fn test_create_new_vault_hashmap_and_store_creation() {
-    use secrets::PasswordBasedCredentialStore;
-    use std::collections::HashMap;
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Test the HashMap creation specifically (line 117)
-    let empty_credentials = HashMap::new();
-    assert!(empty_credentials.is_empty());
-
-    // Test the store creation specifically (line 119)
-    let test_password = "store_creation_test";
-    let store_result = PasswordBasedCredentialStore::new(&empty_credentials, test_password.trim());
-    assert!(store_result.is_ok());
-
-    let store = store_result.unwrap();
-
-    // Test the save operation specifically (line 125)
-    let save_result = store.save_to_file(&cred_path.to_path_buf());
-    assert!(save_result.is_ok());
-
-    assert!(cred_path.exists());
-  }
-
-  #[test]
-  fn test_create_new_vault_success_logging() {
-    // Test that bentley::success can be called (line 127)
-    // We can't easily test the output, but we can verify it doesn't panic
-    bentley::success("vault created successfully");
-
-    // Test various success message formats
-    bentley::success("test message");
-    bentley::success(&format!("formatted message: {}", "value"));
-
-    // Success logging verified by execution completing without panic
-  }
-
-  #[test]
-  fn test_real_create_new_vault_function_success() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Set up test responses for the actual create_new_vault function
-    set_test_prompt_responses(vec!["test_password123".to_string(), "test_password123".to_string()]);
-
-    // Test the ACTUAL create_new_vault function (lines 105-128)
-    let result = create_new_vault(&cred_path);
-    assert!(result.is_ok(), "create_new_vault should succeed with matching passwords");
-    assert_eq!(result.unwrap(), "test_password123");
-    assert!(cred_path.exists(), "Credentials file should be created");
-
-    // Verify the vault works
-    let verify_result = verify_password(&cred_path, "test_password123");
-    assert!(verify_result.is_ok(), "Vault should be verifiable");
-  }
-
-  #[test]
-  fn test_real_create_new_vault_function_empty_password() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Test empty password rejection (lines 108-109)
-    set_test_prompt_responses(vec!["".to_string(), "anything".to_string()]);
-
-    let result = create_new_vault(&cred_path);
-    assert!(result.is_err(), "create_new_vault should fail with empty password");
-    assert!(result.unwrap_err().to_string().contains("master password cannot be empty"));
-  }
-
-  #[test]
-  fn test_real_create_new_vault_function_password_mismatch() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Test password mismatch rejection (lines 113-114)
-    set_test_prompt_responses(vec!["password1".to_string(), "password2".to_string()]);
-
-    let result = create_new_vault(&cred_path);
-    assert!(result.is_err(), "create_new_vault should fail with mismatched passwords");
-    assert!(result.unwrap_err().to_string().contains("passwords do not match"));
-  }
-
-  #[test]
-  fn test_real_create_new_vault_function_whitespace_trimming() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let cred_path = temp_dir.path().join("credentials.enc");
-
-    // Test password trimming (line 128 return)
-    set_test_prompt_responses(vec![
-      "  spaced_password  ".to_string(),
-      "  spaced_password  ".to_string(),
-    ]);
-
-    let result = create_new_vault(&cred_path);
-    assert!(result.is_ok(), "create_new_vault should succeed with spaced passwords");
-    let returned_password = result.unwrap();
-    assert_eq!(returned_password, "spaced_password", "Returned password should be trimmed");
-    assert!(cred_path.exists(), "Credentials file should be created");
-  }
-
-  #[test]
-  fn test_real_create_new_vault_function_directory_creation() {
-    use tempfile::TempDir;
-
-    let temp_dir = TempDir::new().unwrap();
-    let nested_path = temp_dir.path().join("deep").join("nested").join("path");
-    let cred_path = nested_path.join("credentials.enc");
-
-    // Ensure directory doesn't exist initially
-    assert!(!nested_path.exists(), "Nested path should not exist initially");
-
-    // Test directory creation (lines 121-123)
-    set_test_prompt_responses(vec![
-      "dir_test_password".to_string(),
-      "dir_test_password".to_string(),
-    ]);
-
-    let result = create_new_vault(&cred_path);
-    assert!(result.is_ok(), "create_new_vault should create directories");
-    assert!(nested_path.exists(), "Directory should be created (lines 121-123)");
-    assert!(cred_path.exists(), "Credentials file should be created (line 125)");
-  }
+  // All remaining vault creation and password tests moved to crate::encryption::tests
 }
